refactor: 메인 클래스와 패키지 레벨 동일화
