+++ src/test/java/blackjack/domain/deck/CardsTest.java
@@ -1,11 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.deck;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.domain.deck.Card;
-import blackjack.domain.deck.CardShape;
-import blackjack.domain.deck.CardValue;
-import blackjack.domain.deck.Cards;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
