+++ src/test/java/blackjack/domain/deck/DeckTest.java
@@ -1,9 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.deck;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import blackjack.domain.deck.Deck;
 import org.junit.jupiter.api.Test;
 
 public class DeckTest {
