+++ src/test/java/blackjack/domain/participant/DealerTest.java
@@ -1,11 +1,10 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.domain.deck.Card;
 import blackjack.domain.deck.CardShape;
 import blackjack.domain.deck.CardValue;
-import blackjack.domain.participant.Dealer;
 import org.junit.jupiter.api.Test;
 
 public class DealerTest {
