+++ src/test/java/blackjack/domain/participant/UserNameTest.java
@@ -1,8 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import blackjack.domain.participant.UserName;
 import org.junit.jupiter.api.Test;
 
 class UserNameTest {
