+++ src/test/java/blackjack/domain/participant/UsersTest.java
@@ -1,12 +1,10 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.domain.deck.Card;
 import blackjack.domain.deck.CardShape;
 import blackjack.domain.deck.CardValue;
-import blackjack.domain.participant.User;
-import blackjack.domain.participant.Users;
 import java.util.List;
 import org.junit.jupiter.api.Test;
